A local job manager to manage the status of the job on a single node.
