+++ dlrover/python/master/local_master.py
@@ -27,6 +27,7 @@ from dlrover.python.master.elastic_training.rdzv_manager import (
 )
 from dlrover.python.master.master import JobMaster
 from dlrover.python.master.monitor.speed_monitor import SpeedMonitor
+from dlrover.python.master.node.local_job_manager import create_job_manager
 from dlrover.python.master.servicer import create_master_service
 from dlrover.python.master.shard.task_manager import TaskManager
 from dlrover.python.master.stats.job_collector import JobMetricCollector
@@ -37,6 +38,7 @@ class LocalJobMaster(JobMaster):
     def __init__(self, port, args: JobArgs):
         self.speed_monitor = SpeedMonitor()
         self.task_manager = TaskManager(0, self.speed_monitor)
+        self.job_manager = create_job_manager(args, self.speed_monitor)
         elastic_training = RendezvousName.ELASTIC_TRAINING
         self.rdzv_managers: Dict[str, RendezvousManager] = {
             elastic_training: ElasticTrainingRendezvousManager(),
@@ -52,7 +54,7 @@ class LocalJobMaster(JobMaster):
         return create_master_service(
             port,
             self.task_manager,
-            None,
+            self.job_manager,
             self.speed_monitor,
             self.rdzv_managers,
             self.job_metric_collector,
