+++ dlrover/python/tests/test_utils.py
@@ -250,6 +250,9 @@ def mock_k8s_client():
     k8s_client.create_service = mock.MagicMock(  # type: ignore
         return_value=True
     )
+    k8s_client.get_service = mock.MagicMock(  # type: ignore
+        return_value=False
+    )
 
 
 def start_local_master():
